Fix: handle passthrough connections in redirector
Before this fix, the redirector workers could not handle passthrough
connections and forced a 'close' action.

In addition, even if they answer properly ('intercept', which creates a
downloader), it arrives asynchronously. So the initial 'read_client'
hook is not enabled because it may happen before a downloader is
created and the initial data from the client is discarded (losing the
HTTP request).

Signed-off-by: Oriol Arcas <oriol@starflownetworks.com>